// core/enums.ts
export enum UserState {
    start = "start",
    chat = "chat",
    awaiting_project_name = "awaiting_project_name",
    awaiting_project_path = "awaiting_project_path",
    awaiting_repo_url = "awaiting_repo_url",
    awaiting_clone_path = "awaiting_clone_path",
    awaiting_command = "awaiting_command",
    awaiting_chat_message = "awaiting_chat_message",
}